Replace plain DS18 sensor IDs with IDs based on the sensor CRC
This update resolves sensors configuration shifting and mix-up
in the case of one of the sensors being damaged or missing.

Though it introduces another issue, there is a probability of
the sensors ID conflict, in this case, the conflicting sensor just
will not be included in the configuration. This issue is less
critical and doesn't have an influence on reliability unlike the
first.